requirements.txt
django
djangorestframework
PyJWT
python-dotenv
grpcio
grpcio-tools
uvicorn
requests